Add any single unwalkable adjacent cell to the checklist for enclosure, just in case.
With this, the ReplaceStuff mod can build a room with walls into mountain rock, without the corner rock being inaccessible - unmined and unbuilt.

The conflict was - Since ReplaceStuff allows the wall blueprint over the rock, there is no 'region' under the blueprint, so SmarterConstruction didn't have a region to check for blockage. All it needed was to check the single adjacent cell, even if unwalkable, for possibly blocked things, then ReplaceStuff's blueprints are checked for blockage as usual.

Thinking about this more in the future, this mod could probably check adjacent designations like mining and uninstalling that it would block. But that's a bigger change.